+++ R/mapbiomas.R
@@ -4,7 +4,7 @@
 #'
 #' @param dataset A dataset name ("mapbiomas_cover", "mapbiomas_transition", "mapbiomas_irrigation", "mapbiomas_deforestation_regeneration", "mapbiomas_grazing_quality" or "mapbiomas_mining")
 #' @param raw_data A \code{boolean} setting the return of raw or processed data
-#' @param geo_level A \code{string} that defines the geographic level of the data. Can be only "municipality", except "mapbiomas_mining" which accepts more options
+#' @param geo_level A \code{string} that defines the geographic level of the data. Can be only "municipality", except "mapbiomas_mining" which accepts more options ("biome", "indigenous_land")
 #' @param time_period A \code{numeric} indicating what years will the data be loaded. Can be only "all".
 #' @param language A \code{string} that indicates in which language the data will be returned. Currently, only Portuguese ("pt") and English ("eng") are supported.
 #' @param time_id A \code{string} that indicates the time criteria for the data loaded. Can be "year" or "month". Defaults to year.
@@ -14,22 +14,24 @@
 #' @examples
 #' \dontrun{
 #' # download treated data from mapbiomas_grazing_quality
-#' treated_mapbiomas_grazing <- load_mapbiomas(dataset = "mapbiomas_grazing_quality",
-#'                             raw_data = FALSE, geo_level = "municipality",
-#'                             time_period = "all", language = "pt")
+#' treated_mapbiomas_grazing <- load_mapbiomas(
+#'   dataset = "mapbiomas_grazing_quality",
+#'   raw_data = FALSE, geo_level = "municipality",
+#'   time_period = "all", language = "pt"
+#' )
 #' }
 #'
 #' @importFrom magrittr %>%
 #' @export
 
 
-load_mapbiomas = function(dataset = NULL,raw_data=NULL,geo_level = 'municipality',
-               time_period='all',language = 'eng',time_id = 'year',cover_level = 1){
+load_mapbiomas <- function(dataset = NULL, raw_data = NULL, geo_level = "municipality",
+                           time_period = "all", language = "eng", time_id = "year", cover_level = 1) {
 
   ## To-Do:
-    ## Add Support to Transition Info Download at the State and Biome Level
-    ## Add Support to Covering Info Download at the State and Biome Level
-    ## Include Support for raw raster download
+  ## Add Support to Transition Info Download at the State and Biome Level
+  ## Add Support to Covering Info Download at the State and Biome Level
+  ## Include Support for raw raster download
 
 
   ###########################
@@ -37,7 +39,7 @@ load_mapbiomas = function(dataset = NULL,raw_data=NULL,geo_level = 'municipality
   ###########################
   survey <- link <- x1985 <- x2019 <- NULL
   ano <- bioma <- category <- cidade <- city <- class_id <- country <- estado <- feature_id <- group <- terra_indigena <- NULL
-  id <- indigenous_lands <- level_2 <- level_3 <- name_pt_br <- pais <-x2020 <- NULL
+  id <- indigenous_land <- level_2 <- level_3 <- name_pt_br <- pais <- x2020 <- NULL
   territory_id <- municipality <- state <- year <- value <- NULL
   x1985_to_1986 <- x2018_to_2019 <- x1988 <- x2017 <- x2000 <- x2010 <- x2018 <- biome <- level_1 <- NULL
 
@@ -46,450 +48,162 @@ load_mapbiomas = function(dataset = NULL,raw_data=NULL,geo_level = 'municipality
   ## Define Basic Parameters ##
   #############################
 
-  param=list()
-  param$dataset = dataset
-  param$geo_level = geo_level
-  param$time_period = time_period
-  param$language = language
-  param$time_id = time_id
-  param$raw_data = raw_data
+  param <- list()
+  param$dataset <- dataset
+  param$geo_level <- geo_level
+  param$time_period <- time_period
+  param$language <- language
+  param$time_id <- time_id
+  param$raw_data <- raw_data
 
-  param$survey_name = datasets_link() %>%
+  param$survey_name <- datasets_link() %>%
     dplyr::filter(dataset == param$dataset) %>%
     dplyr::select(survey) %>%
     unlist()
 
-  param$url = datasets_link() %>%
+  param$url <- datasets_link() %>%
     dplyr::filter(dataset == param$dataset) %>%
     dplyr::select(link) %>%
     unlist()
 
   ## Dataset
 
-  if (is.null(param$dataset)){stop('Missing Dataset!')}
-  if (is.null(param$raw_data)){stop('Missing TRUE/FALSE for Raw Data')}
+  if (is.null(param$dataset)) {
+    stop("Missing Dataset!")
+  }
+  if (is.null(param$raw_data)) {
+    stop("Missing TRUE/FALSE for Raw Data")
+  }
 
+  sheets <- tibble::tribble(
+    ~dataset, ~geo_level, ~sheet,
+    "mapbiomas_cover", "any", "LAND COVER - BIOMAS e UF",
+    "mapbiomas_transition", "any", "BD_TRANSICAO_BIOMA-UF",
+    "mapbiomas_deforestation_regeneration", "any", "BD Colecao 5.0(h) - Hectares",
+    "mapbiomas_irrigation", "any", "BD_IRRIGACAO",
+    "mapbiomas_grazing_quality", "any", "BD_Qualidade",
+    "mapbiomas_mining", "country", "BR",
+    "mapbiomas_mining", "state", "UF",
+    "mapbiomas_mining", "biome", "BIOME",
+    "mapbiomas_mining", "municipality", "MUN",
+    "mapbiomas_mining", "indigenous_land", "TI"
+  )
+
+  sheet <- sheets %>%
+    dplyr::filter(
+      dataset == param$dataset,
+      geo_level %in% c(param$geo_level, "any")
+    ) %>%
+    dplyr::select(sheet) %>%
+    unlist()
 
   #################
   ## Downloading ##
   #################
 
-  dat = external_download(dataset = param$dataset,source = 'mapbiomas', geo_level = param$geo_level)
+  dat <- external_download(
+    dataset = param$dataset,
+    source = "mapbiomas",
+    geo_level = param$geo_level,
+    sheet = sheet
+  )
 
-  dat = dat %>%
+  dat <- dat %>%
     janitor::clean_names() %>%
     tibble::as_tibble() %>%
-    dplyr::mutate_if(is.character,function(var){stringi::stri_trans_general(str=var,id="Latin-ASCII")})
+    dplyr::mutate_if(is.character, function(var) {
+      stringi::stri_trans_general(str = var, id = "Latin-ASCII")
+    })
 
-  if (raw_data == TRUE){return(dat)}
+  if (raw_data == TRUE) {
+    return(dat)
+  }
 
   ######################
   ## Data Engineering ##
   ######################
 
-  if (param$dataset == 'mapbiomas_cover'){
+  ## Create Longer Data - Years as a Variable
 
-    ## Create Longer Data - Years as a Variable
+  dat <- dat %>%
+    tidyr::pivot_longer(
+      cols = dplyr::starts_with("x"),
+      names_to = "year",
+      names_prefix = "x",
+      values_to = "value"
+    )
 
-    dat = dat %>%
-      tidyr::pivot_longer(
-        cols = x1985:x2019,
-        names_to = 'year',
-        names_prefix = 'x',
-        values_to = 'value'
-      )
-
-    ## Aggregating by Cover Level
-
-    if (cover_level == 0){dat$cover_level = dat$level_0}
-    if (cover_level == 1){dat$cover_level = dat$level_1}
-    if (cover_level == 2){dat$cover_level = dat$level_2}
-    if (cover_level == 3){dat$cover_level = dat$level_3}
-    if (cover_level == 4){dat$cover_level = dat$level_4}
-
-    dat = dat %>%
-      tidyr::pivot_wider(id_cols = c(territory_id,municipality,state,year),
-                         names_from = cover_level,
-                         values_from = value,
-                         values_fn = sum,
-                         values_fill = NA) %>%
-      janitor::clean_names()
-
-    ## Adjusting Geo Level Names
-
-    ## Translating Names
-
-    ## Add Labels
-
-    ## Return Data
-
-    return(dat)
-
-  }
-
-  if (param$dataset == 'mapbiomas_transition'){
-
-    ## Create Longer Data - Years as a Variable
-
-    dat = dat %>%
-      tidyr::pivot_longer(
-        cols = x1985_to_1986:x2018_to_2019,
-        names_to = 'transition_year',
-        names_prefix = 'x',
-        values_to = 'value'
-      )
+  if (param$dataset %in% c("mapbiomas_cover", "mapbiomas_transition")) {
 
     ## Aggregating by Cover Level
 
-    # if (cover_level == 0){dat$cover_level = dat$level_0}
-    # if (cover_level == 1){dat$cover_level = dat$level_1}
-    # if (cover_level == 2){dat$cover_level = dat$level_2}
-    # if (cover_level == 3){dat$cover_level = dat$level_3}
-    # if (cover_level == 4){dat$cover_level = dat$level_4}
-    #
-    # dat = dat %>%
-    #   tidyr::pivot_wider(id_cols = c(territory_id,municipality,state,year),
-    #                      names_from = cover_level,
-    #                      values_from = value,
-    #                      values_fn = sum,
-    #                      values_fill = NA) %>%
-    #   janitor::clean_names()
+    if (cover_level == 0) {
+      dat$cover_level <- dat$level_0
+    }
+    if (cover_level == 1) {
+      dat$cover_level <- dat$level_1
+    }
+    if (cover_level == 2) {
+      dat$cover_level <- dat$level_2
+    }
+    if (cover_level == 3) {
+      dat$cover_level <- dat$level_3
+    }
+    if (cover_level == 4) {
+      dat$cover_level <- dat$level_4
+    }
 
-    ## Adjusting Geo Level Names
-
-    ## Translating Names
-
-    ## Add Labels
-
-    ## Return Data
-
-    return(dat)
-
-
-
-
-  }
-
-  if (param$dataset == 'mapbiomas_deforestation_regeneration'){
-
-    ## Create Longer Data - Years as a Variable
-
-    dat = dat %>%
-      tidyr::pivot_longer(
-        cols = x1988:x2017,
-        names_to = 'year',
-        names_prefix = 'x',
-        values_to = 'value'
-      )
-
-    ## Return Data
-
-    return(dat)
-
-  }
-
-  if (param$dataset == 'mapbiomas_irrigation'){
-
-    ## Create Longer Data - Years as a Variable
-
-    dat = dat %>%
-      tidyr::pivot_longer(
-        cols = x2000:x2019,
-        names_to = 'year',
-        names_prefix = 'x',
-        values_to = 'value'
-      )
-
-    ## Return Data
-
-    return(dat)
-
-
-
-  }
-
-  if (param$dataset == 'mapbiomas_grazing_quality'){
-
-    dat = dat %>%
-      tidyr::pivot_longer(
-        cols = x2010:x2018,
-        names_to = 'year',
-        names_prefix = 'x',
-        values_to = 'value'
-      )
-
-    dat = dat %>%
-      tidyr::pivot_wider(id_cols = c(biome,state,year),
-                         names_from = level_1,
-                         values_from = value,
-                         values_fn = sum,
-                         values_fill = NA) %>%
+    dat <- dat %>%
+      tidyr::pivot_wider(
+        id_cols = dplyr::any_of(c(
+          territory_id, municipality, state, year,
+          biome, state, year
+        )),
+        names_from = cover_level,
+        values_from = value,
+        values_fn = sum,
+        values_fill = NA
+      ) %>%
       janitor::clean_names()
 
-    ## Return Data
-
     return(dat)
-
-
-
-
   }
 
-  if(param$dataset == "mapbiomas_mining" & param$geo_level == "country" & param$language == "pt"){
+  dat <- dat %>%
+    dplyr::select(-category)
 
-      ## Create Longer Data - Years as a Variable
+  #################
+  ## Translation ##
+  #################
 
-      dat = dat %>%
-        tidyr::pivot_longer(
-          cols = x1985:x2020,
-          names_to = 'year',
-          names_prefix = 'x',
-          values_to = 'value'
-        )
+  dat_mod <- dat %>%
+    dplyr::rename_with(dplyr::recode,
+      "name_pt_br" = param$geo_level
+    )
 
-      dat = dat %>%
-        dplyr::select(-c(category))
+  dat_mod <- dat_mod %>%
+    dplyr::relocate(dplyr::any_of(c("year", "state", "city", "biome", "indigenous_land")))
 
-      dat = dat %>%
-      dplyr::rename(id = feature_id,
-                    pais = name_pt_br,
-                    id_classe = class_id,
-                    grupo = group,
-                    tipo_mineracao = level_1,
-                    categoria_mineracao = level_2,
-                    produtos_mineracao = level_3,
-                    ano = year,
-                    valor = value)
-
-      dat = dat %>%
-        dplyr::relocate(ano, pais, id)
-  }
-
-  if(param$dataset == "mapbiomas_mining" & param$geo_level == "country" & param$language == "eng"){
-
-    ## Create Longer Data - Years as a Variable
-
-    dat = dat %>%
-      tidyr::pivot_longer(
-        cols = x1985:x2020,
-        names_to = 'year',
-        names_prefix = 'x',
-        values_to = 'value'
+  if (param$language == "pt") {
+    dat_mod <- dat_mod %>%
+      dplyr::rename_with(dplyr::recode,
+        "feature_id" = "id",
+        "class_id" = "id_classe",
+        "group" = "grupo",
+        "year" = "ano",
+        "value" = "valor",
+        "state" = "estado",
+        "municipality" = "municipio",
+        "biome" = "bioma",
+        "indigenous_land" = "terra_indigena"
       )
-
-    dat = dat %>%
-      dplyr::select(-c(category))
-
-    dat = dat %>%
-      dplyr::rename(country = name_pt_br,
-                    mining_type = level_1,
-                    mining_category = level_2,
-                    mining_products = level_3)
-
-    dat = dat %>%
-      dplyr::relocate(year, country, feature_id)
   }
 
-  if(param$dataset == "mapbiomas_mining" & param$geo_level == "state" & param$language == "pt"){
-
-    dat = dat %>%
-      tidyr::pivot_longer(
-        cols = x1985:x2020,
-        names_to = 'year',
-        names_prefix = 'x',
-        values_to = 'value'
-      )
-
-    dat = dat %>%
-      dplyr::select(-c(category)) %>%
-      dplyr::rename(estado = name_pt_br,
-                    tipo_mineracao = level_1,
-                    categoria_mineracao = level_2,
-                    produtos_mineracao = level_3,
-                    id = feature_id,
-                    id_classe = class_id,
-                    ano = year,
-                    valor = value)
-
-    dat = dat %>%
-      dplyr::relocate(ano, estado, id)
-  }
-
-  if(param$dataset == "mapbiomas_mining" & param$geo_level == "state" & param$language == "eng"){
-
-    ## Create Longer Data - Years as a Variable
-
-    dat = dat %>%
-      tidyr::pivot_longer(
-        cols = x1985:x2020,
-        names_to = 'year',
-        names_prefix = 'x',
-        values_to = 'value'
-      )
-
-    dat = dat %>%
-      dplyr::select(-c(category))
-
-    dat = dat %>%
-      dplyr::rename(state = name_pt_br,
-                    mining_type = level_1,
-                    mining_category = level_2,
-                    mining_products = level_3)
-
-    dat = dat %>%
-      dplyr::relocate(year, country, feature_id)
-  }
-
-  if(param$dataset == "mapbiomas_mining" & param$geo_level == "municipality" & param$language == "pt"){
-
-    dat = dat %>%
-      tidyr::pivot_longer(
-                 cols = x1985:x2020,
-                 names_to = 'year',
-                 names_prefix = 'x',
-                 values_to = 'value')
-
-    dat = dat %>%
-      dplyr::select(-c(category))
-
-    dat = dat %>%
-      dplyr::rename(id = feature_id,
-                    municipio = city,
-                    id_classe = class_id,
-                    grupo = group,
-                    tipo_mineracao = level_1,
-                    categoria_mineracao = level_2,
-                    produtos_mineracao = level_3,
-                    ano = year,
-                    valor = value,
-                    estado = state)
-
-    dat = dat %>%
-      dplyr::relocate(ano, cidade, estado, id)
-  }
-
-  if(param$dataset == "mapbiomas_mining" & param$geo_level == "municipality" & param$language == "eng"){
-
-    dat = dat %>%
-      tidyr::pivot_longer(
-        cols = x1985:x2020,
-        names_to = 'year',
-        names_prefix = 'x',
-        values_to = 'value'
-      )
-
-    dat = dat %>%
-      dplyr::select(-c(category))
-
-    dat = dat %>%
-      dplyr::rename(mining_type = level_1,
-                    mining_category = level_2,
-                    mining_products = level_3)
-
-    dat = dat %>%
-      dplyr::relocate(year, city, state, feature_id)
-  }
-
-  if(param$dataset == "mapbiomas_mining" & param$geo_level == "biome" & param$language == "pt"){
-
-    dat = dat %>%
-      tidyr::pivot_longer(
-        cols = x1985:x2020,
-        names_to = 'year',
-        names_prefix = 'x',
-        values_to = 'value')
-
-    dat = dat %>%
-      dplyr::select(-c(category))
-
-    dat = dat %>%
-      dplyr::rename(id = feature_id,
-                    bioma = name_pt_br,
-                    id_classe = class_id,
-                    grupo = group,
-                    tipo_mineracao = level_1,
-                    categoria_mineracao = level_2,
-                    produtos_mineracao = level_3,
-                    ano = year,
-                    valor = value)
-
-    dat = dat %>%
-      dplyr::relocate(ano, bioma, id)
-
-  }
-
-  if(param$dataset == "mapbiomas_mining" & param$geo_level == "biome" & param$language == "eng"){
-
-    dat = dat %>%
-      tidyr::pivot_longer(
-        cols = x1985:x2020,
-        names_to = 'year',
-        names_prefix = 'x',
-        values_to = 'value')
-
-    dat = dat %>%
-      dplyr::select(-c(category))
-
-    dat = dat %>%
-      dplyr::rename(mining_type = level_1,
-                    mining_category = level_2,
-                    mining_products = level_3,
-                    biome = name_pt_br)
-
-    dat = dat %>%
-      dplyr::relocate(year, biome, feature_id)
-  }
-
-  if(param$dataset == "mapbiomas_mining" & param$geo_level == "indigenous" & language == "pt"){
-
-    dat = dat %>%
-      tidyr::pivot_longer(
-        cols = x1985:x2020,
-        names_to = 'year',
-        names_prefix = 'x',
-        values_to = 'value')
-
-    dat = dat %>%
-      dplyr::select(-c(category))
-
-    dat = dat %>%
-      dplyr::rename(id = feature_id,
-                    terra_indigena = name_pt_br,
-                    id_classe = class_id,
-                    grupo = group,
-                    tipo_mineracao = level_1,
-                    categoria_mineracao = level_2,
-                    produtos_mineracao = level_3,
-                    ano = year,
-                    valor = value)
-
-    dat = dat %>%
-      dplyr::relocate(ano, terra_indigena, id)
-  }
-
-  if(param$dataset == "mapbiomas_mining" & param$geo_level == "indigenous" & language == "eng"){
-
-    dat = dat %>%
-      tidyr::pivot_longer(
-        cols = x1985:x2020,
-        names_to = 'year',
-        names_prefix = 'x',
-        values_to = 'value')
-
-    dat = dat %>%
-      dplyr::select(-c(category))
-
-    dat = dat %>%
-      dplyr::rename(mining_type = level_1,
-                    mining_category = level_2,
-                    mining_products = level_3,
-                    indigenous_lands = name_pt_br)
-
-    dat = dat %>%
-      dplyr::relocate(year, indigenous_lands, feature_id)
-  }
-
-  return(dat)
 
 
+  ####################
+  ## Returning Data ##
+  ####################
+
+  return(dat_mod)
 }